fix booking table to match required field namings
